Fix the mouse associated with the zoom in, zoom out feature

+++ GameBoardEditor/src/main/java/heron/gameboardeditor/EditingScreenController.java
@@ -68,7 +68,7 @@ public class EditingScreenController {
     private VBox boardParentVBox;
     private GridBoardUI gridBoard;
     private UndoRedoHandler undoRedoHandler;
-    private double tileSize;
+    private int tileSize;
     
     @FXML
     private void initialize() {
